rename 'types.py' module to 'test_types.py'
